requirements.txt

Flask==2.3.3
Flask-SQLAlchemy==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.0
pymysql==1.1.0
